Add secure coding support for migration
Apparently, the migration support should use Secure Coding, or
else it will fail miserably and possibly crash later in the
function for users. Now it will use Secure Coding import, falling
back on insecure coding. And it will check if the decode failed
entirely, returning nil if the incoming object type is NSData.

This was otherwise failing and crashing where it treated the
NSData like an NSDictionary.

Signed-off-by: Christopher Snowhill <kode54@gmail.com>